# doc/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "",
    "version" : "1.0.0",
    "title" : "Micro serviço de Carrinho de Compras",
    "contact" : {
      "email" : "weto.jc@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:3000",
  "basePath" : "/api/v1",
  "tags" : [ {
    "name" : "carrinhocompra",
    "description" : "Chamadas de serviço de Carrinho de compras"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/carrinhocompra" : {
      "get" : {
        "tags" : [ "carrinhocompra" ],
        "summary" : "listagem de todos os produtos no carrinho de compra",
        "description" : "Listar de todos os produtos no carrinho de compra.\n",
        "operationId" : "listCarrinhocompra",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Buscar todos os Produtos no carrinho de compra.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CarrinhoCompra"
              }
            }
          }
        }
      }
    },
    "/carrinhocompra/adicionar/{idProduto}" : {
      "post" : {
        "tags" : [ "carrinhocompra" ],
        "summary" : "Adicionar produto no carrinho de compras",
        "description" : "Serviço para adicionar produto no carrinho de compras: \n",
        "operationId" : "AddCarrinhoCompra",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "idProduto",
          "description" : "Identificação do Produto",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Adicionar o Id do produto no carrinho de compras.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CarrinhoCompra"
              }
            }
          }
        }
      }
    },
    "/carrinhocompra/remover/{idProduto}" : {
      "delete" : {
        "tags" : [ "carrinhocompra" ],
        "summary" : "Remover o produto no carrinho de compra",
        "description" : "Serviço para remover produto no carrinho de compra: \n",
        "operationId" : "RemoveCarrinhoCompra",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "idProduto",
          "description" : "Identificação do Produto",
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "Remover o Id do produto no carrinho de compras.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CarrinhoCompra"
              }
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "CarrinhoCompra" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "uuid",
          "example" : 1
        }
      },
      "example" : {
        "id" : 1
      }
    }
  }
}